+++ lib/marskal_doc.rb
@@ -1,0 +1,13 @@
+require "marskal_rdoc/examples"
+require "marskal_rdoc/standards"
+
+
+##
+# MarskalRdoc. This module will have examples and suggest standards for implementing source code documentation for both standard <tt>RDoc</tt>
+# markup syntax and the extension to RDoc via the {yard gem}[https://rubygems.org/gems/yard] .
+#
+#:include:marskal_rdoc/includes/reference_links.rb
+module MarskalRdoc
+
+
+end