+++ lib/marskal_rdoc/examples/rdoc/links.rb
@@ -5,12 +5,7 @@ module MarskalRdoc
       ##
       # Links. This module should have a method for most of the common links for RDoc
       #
-      # <b>Much of the information was provided from the following links:</b>
-      #
-      # - http://www.mikeperham.com/wp-content/uploads/2010/12/rdoc.html
-      # - http://jan.varwig.org/wp-content/uploads/2006/09/Rdoc%20Cheat%20Sheet.pdf
-      # - http://docs.seattlerb.org/rdoc/RDoc/Markup.html
-      #
+      #:include:/../../includes/reference_links.rb
       module Links
         ##
         # This demonstrates RDoc automatically links your classes and file names when included in the documentation