+++ lib/marskal_rdoc/standards.rb
@@ -1,0 +1,14 @@
+require_relative 'standards/rdoc'
+
+module MarskalRdoc
+
+  ##
+  # Standards. This module will have practical examples of a standard way to document ruby objects including modules, classes, variables and more
+  #
+  #:include:includes/reference_links.rb
+  module Standards
+
+
+  end
+
+end