+++ lib/marskal_rdoc/standards/rdoc.rb
@@ -1,0 +1,17 @@
+module MarskalRdoc
+  module Standards
+
+    ##
+    # Rdoc. Rdoc
+    #
+    # <b>This module will show standards for documenting using pure RDoc syntax and tags only.</b>
+    #
+    module Rdoc
+
+      # View Source to see how this was documented
+      SAMPLE_CONSTANT_VARIABLE = 'Example of Constant'
+
+    end
+
+  end
+end